Backup DB and start empty if migrations fail
In case the migrations fail, we back up the database to a file including
the current timestamp and reconnect to an empty one.

This allows the application to startup and thus improves resilience.
It also allows us to make changes to our migration files.

Note on the `BoxFuture`: This is required to enable recursion of async
functions.